scripts/requirements.txt
crcmod
cryptography
hidapi
